feat: vite custom plugin 코드 분리
